requirements.txt
torch
torchvision
torchaudio
dvc
pandas
pillow
opencv-python
scikit-learn
matplotlib
tqdm
jupyter
qdrant-client
fastapi
uvicorn
python-multipart